pyproject.toml
[tool.poetry]
name = "atomic-habits-project"
version = "0.1.0"
description = ""
authors = ["Uliana Semyonova <i@ulianasemyonova.ru>"]
readme = "README.md"
packages = [{include = "atomic_habits_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
redis = "^5.0.1"
djangorestframework = "^3.14.0"
ipython = "^8.16.1"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
django-cors-headers = "^4.3.0"
drf-spectacular = "^0.26.5"
djangorestframework-simplejwt = "^5.3.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
coverage = "^7.3.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
